pyproject.toml
[project]
name = "autonomous-incident-response-env"
version = "1.0.0"
description = "Autonomous Incident Response Benchmark Environment for evaluating AI agents on structured runbook execution and decision-making"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "openai",
    "pydantic",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
